remove bug cast double to int in move_player

+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: watanabekoji <watanabekoji@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/02 22:48:56 by kojwatan          #+#    #+#             */
-/*   Updated: 2024/07/09 23:41:52 by watanabekoj      ###   ########.fr       */
+/*   Updated: 2024/07/10 12:27:55 by watanabekoj      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -114,8 +114,8 @@ typedef struct s_player
 {
 	int		pos_x;
 	int		pos_y;
-	float	dir_x;
-	float	dir_y;
+	double	dir_x;
+	double	dir_y;
 }	t_player;
 
 typedef struct s_info
@@ -192,8 +192,8 @@ bool	is_next_to_blank(t_node **map, int x, int y);
 //-----------------------
 void	rotate_camera(t_vars *vars, int keycode);
 void	rotate_player(t_player *player, int keycode, int count);
-float	new_dir_x(t_player player, double rad);
-float	new_dir_y(t_player player, double rad);
+double	new_dir_x(t_player player, double rad);
+double	new_dir_y(t_player player, double rad);
 
 //-----------------------
 //			  move_player.c
